Revival of tracepoints by RWX -> W^X remapping at loader
Note that tracepoints that must be added in runtime cannot have W^X. In this case, --runtime-tracepoint argument can be given to abandon W^X and resort to RWX-mapped kernel binary.
Also, it is hard to know in advance whether or not we would need RWX or W^X. This is because kernel loading is done in advance compared to argv parsing later on at the loader. This is why a re-mapping technique is used.